restructured files for part 2, completed part 2
