+++ part-2-space/server.js
@@ -2,6 +2,8 @@
 const express = require('express');
 const app = express();
 
+const marsMissions = require('./models/marsMissions.js')
+
 // run `npm install` to install dependencies in package.json
 
 // * Your mission is to complete the app
@@ -22,51 +24,31 @@ const port = 3000;
 
 // DATA - put into marsMissions.js file inside of a models folder, for module.exports
 // remember to require it in the server
-const marsMissions = [
-  {
-    name: "Curiosity",
-    launchDate: "26 Nov 2011",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Opportunity",
-    launchDate: "8 Jul 2003",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Spirit",
-    launchDate: "10 Jun 2003",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Sojourner",
-    launchDate: "4 Dec 1996",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Rosetta",
-    launchDate: "2 Mar 2004",
-    operator: "ESA",
-    missionType: "Gravity Assist",
-    img: ""
-  }
-];
+
+app.get('/', (req,res) => {
+  res.send('connection established')
+})
+
 
 // INDEX Route
 // send data to 'missions/index.ejs' view
 // the view should display just the names of each mission
 
+app.get('/index', (req,res) => {
+  res.render('index.ejs', {
+    marsMissions: marsMissions[req.params.show]
+  });
+})
+
 // SHOW Route
 // send data to 'missions/show.ejs' view
 // the view should display all the data for a single mission
+
+app.get('/index/:show', (req,res) => {
+  res.render('show.ejs', {
+    marsMissions: marsMissions[req.params.show]
+  });
+})
 
 
 // LISTENER
