+++ part-2-space/views/index.ejs
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<meta charset="utf-8">
+	<title></title>
+</head>
+<body>
+	<h3>Index of Mars Missions:</h3>
+	</br>
+
+	<a href="http://localhost:3000/index/0">Curiosity</a> </br>
+	<a href="http://localhost:3000/index/1">Opportunity</a> </br>
+	<a href="http://localhost:3000/index/2">Spirit</a> </br>
+	<a href="http://localhost:3000/index/3">Sojourner</a> </br>
+	<a href="http://localhost:3000/index/4">Rosetta</a> </br>
+
+</body>
+</html>
